fix(gallery): refactor code
