+++ src/lib/viewers/gallery/GalleryGridShared.tsx
@@ -1,0 +1,73 @@
+import React from 'react';
+import { replacePlaceholders } from '../../util';
+import { HighResRenderTask } from './HighResThumbnailStore';
+
+export interface GalleryThumbnail {
+    init: () => Promise<unknown>;
+    getImageFromCache: (itemIndex: number) => { image?: HTMLImageElement; inProgress: boolean } | null | undefined;
+    createThumbnailImage: (
+        itemIndex: number,
+        options: { createImgTag: boolean; thumbMaxWidth: number },
+    ) => Promise<HTMLImageElement | null>;
+    renderPageImage: (pageNum: number, options: { thumbMaxWidth: number }) => HighResRenderTask;
+    /** First-page width:height ratio, populated by init(). Used to size placeholders to the real page shape. */
+    pageRatio?: number;
+}
+
+export interface TileProps {
+    pageNum: number;
+    isFocused: boolean;
+    ariaColIndex?: number;
+    imageSrc?: string;
+    onClick: (pageNum: number) => void;
+    onFocus: (pageNum: number) => void;
+    pageRatio?: number | null;
+    role: 'option' | 'gridcell';
+    width?: number;
+}
+
+export const GalleryTile = React.memo(function GalleryTile({
+    pageNum,
+    isFocused,
+    ariaColIndex,
+    imageSrc,
+    onClick,
+    onFocus,
+    pageRatio,
+    role,
+    width,
+}: TileProps): JSX.Element {
+    const ratio = pageRatio && Number.isFinite(pageRatio) && pageRatio > 0 ? pageRatio : null;
+    const tileStyle = {
+        ...(ratio ? { aspectRatio: String(ratio) } : undefined),
+        ...(width != null ? { width } : undefined),
+    };
+    const contentStyle = ratio ? { height: '100%' } : undefined;
+    const placeholderStyle = ratio ? { ...contentStyle, paddingTop: 0 } : undefined;
+
+    return (
+        // eslint-disable-next-line jsx-a11y/click-events-have-key-events
+        <div
+            aria-colindex={ariaColIndex}
+            aria-label={replacePlaceholders(__('page_gallery_tile'), [String(pageNum)])}
+            aria-selected={isFocused}
+            className={`bp-gallery-tile${isFocused ? ' bp-gallery-tile--selected' : ''}`}
+            data-page={pageNum}
+            data-resin-target="galleryTile"
+            onClick={() => onClick(pageNum)}
+            onFocus={() => onFocus(pageNum)}
+            role={role}
+            style={tileStyle}
+            tabIndex={isFocused ? 0 : -1}
+        >
+            <span aria-hidden="true" className="bp-gallery-tile-badge">
+                {pageNum}
+            </span>
+            {imageSrc ? (
+                <img alt="" src={imageSrc} style={contentStyle} />
+            ) : (
+                <span className="bp-gallery-tile-placeholder" style={placeholderStyle} />
+            )}
+        </div>
+    );
+});
